// rmp_ui_test/src/components/Table/src/types/table.ts
export declare type SortOrder = 'ascend' | 'descend';
// ...
export interface SorterResult {
  column: ColumnProps<any>;
  order: SortOrder;
  field: string;
  columnKey: string;
}

export interface FetchParams {
  searchInfo?: Recordable;
  pagination?: { current?: number; pageSize?: number };
  sortInfo?: Recordable;
  filterInfo?: Recordable;
}
// ...
export interface BasicTableProps<T = any> extends BasicTableExtraProps {
// ...
  onFetchSuccess?: (params: { items: Recordable[]; total: number }) => void;
  onFetchError?: (err: any) => void;
// ...
}